requirements.txt
discord.py[voice]
googletrans==3.1.0a0
black
sqlalchemy
Pillow==9.*.*
PyNaCl
requests
flask
beautifulsoup4==4.9.1
certifi==2022.12.7
chardet==3.0.4
idna==2.10
soupsieve==2.0.1
urllib3==1.26.14
GitPython
python-dotenv
youtube_dl
mojang==1.1.0
moviepy
uuid
dpytest
pytest
